protecting posts w/ authorization

+++ backend/routes/posts.js
@@ -57,7 +57,7 @@ router.post("",
 router.put("/:id",
   checkAuth,
   multer({storage: storage}).single("image"),
-  (req,res,next) => {
+  (req, res, next) => {
   let imagePath = req.body.imagePath;
   if (req.file) {
     const url = req.protocol + '://' + req.get("host");
@@ -70,9 +70,13 @@ router.put("/:id",
     imagePath: imagePath
   })
   // console.log(post);
-  Post.updateOne({_id: req.params.id}, post).then(result => {
-    //console.log(result);
-    res.status(200).json({message: 'Update successful'});
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+    // console.log(result);
+    if (result.modifiedCount > 0) {
+      res.status(200).json({message: 'Update successful'});
+    } else {
+      res.status(401).json({message: 'Not Authorized'});
+    }
   });
 });
 
@@ -111,10 +115,14 @@ router.get("/:id", (req, res, next) => {
 })
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then((result) => {
-    console.log(result);
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then((result) => {
+    // console.log(result);
+    if (result.deletedCount > 0) {
+      res.status(200).json({ message: "Post deleted" });
+    } else {
+      res.status(401).json({message: "Delete Not Authorized"});
+    }
   });
-  res.status(200).json({ message: "Post deleted" });
 });
 
 module.exports = router;
